Move collider constuctors to `constructor` module
